str_sql -> core
